@@ -1,59 +1,54 @@
 package com.avogine;
 
-import java.lang.invoke.*;
 import java.util.concurrent.*;
 
-import org.slf4j.*;
-
 import com.avogine.game.*;
-import com.avogine.io.*;
+import com.avogine.io.Window;
+import com.avogine.logging.AvoLog;
+import com.avogine.util.FrameProfiler;
 
 /**
  * This is the primary entry point into running a game.
  * <p>
  * To kick off the game loop, create a new {@link Avogine} and supply it with a {@link Window} to render
  * to and a {@link Game} to run, then call {@link #start()}.
- * @author Dominus
  *
  */
 public class Avogine implements Runnable {
 
-	private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass().getSimpleName());
-
 	/**
-	 * TODO This needs to be customizable/read from a property.
-	 * Target frames per second. Controls how often the screen is rendered in a single second.
+	 * When sleeping the game loop thread per frame, the system will issue {@code Thread.sleep(1)} requests for as long as the
+	 * remaining frame time is greater than this value. Once it's below this value it will go into a busy wait until 
+	 * the frame time is up.
+	 * </p>
+	 * Certain platforms may have higher error rates on their {@code Thread.sleep()} precision, and this value could be
+	 * increased to better suit them. Although a higher value will mean more busy waiting leading to more CPU cycling.
 	 */
-	public static final int TARGET_FPS = 144;
-
-	/**
-	 * Target updates per second. Controls how often game logic is run in a single second.
-	 */
-	public static final int TARGET_UPS = 60;
+	private static final long SLEEP_PRECISION = TimeUnit.MILLISECONDS.toNanos(2);
 
 	private final Thread gameLoopThread;
+	private final FrameProfiler profiler;
 	
 	private final Window window;
 	private final Game game;
 	
 	private final Timer timer;
-	private final Input input;
 	
-	private float frameTime;
-	private int fps;
-	private float sleepRemainder;
-		
+	private double updateInterval;
+	private double updateAccumulator;
+	
 	/**
 	 * @param window The primary game {@link Window} to display to
 	 * @param game An implementation of {@link Game} that contains the main game logic
 	 */
 	public Avogine(Window window, Game game) {
 		gameLoopThread = new Thread(this, "GAME_LOOP_THREAD");
+		profiler = FrameProfiler.NO_OP;
+		
 		this.window = window;
 		this.game = game;
+		
 		timer = new Timer();
-		// TODO Add an additional constructor for custom Input implementations as well, maybe even Timer? But idc right now.
-		input = new Input();
 	}
 	
 	/**
@@ -66,7 +61,7 @@ public class Avogine implements Runnable {
 		String osName = System.getProperty("os.name");
 		if (osName.contains("Mac")) {
 			System.setProperty("java.awt.headless", "true");
-			logger.info("Setting java.awt.headless to true");
+			AvoLog.log().info("Setting java.awt.headless to true");
 			gameLoopThread.run();
 		} else {
 			gameLoopThread.start();
@@ -79,123 +74,119 @@ public class Avogine implements Runnable {
 			init();
 			loop();
 		} catch (Exception e) {
-			logger.error("An error occurred.", e);
+			AvoLog.log().error("An error occurred.", e);
 		} finally {
 			cleanup();
 		}
 	}
 	
 	private void init() {
-		window.init(input);
-		game.init(window);
+		window.init(game.getGLFWConfig(), game.getInputConfig());
+		game.baseInit(window);
 		timer.init();
-		
-		// XXX This sort of works, but only when the window is resized, not while it's being moved/held
-//		GLFW.glfwSetWindowRefreshCallback(window.getId(), (window) -> {
-//			logger.info("Refreshing");
-//			
-//			update(0.045f);
-//			
-//			fps++;
-//			if (frameTime >= 1.0f) {
-//				window.setFps(fps);
-//				frameTime = 0;
-//				fps = 0;
-//			}
-//			stage.render(window);
-//			
-//			GLFW.glfwSwapBuffers(window);
-//		});
-		
 	}
 	
-	/**
-	 * Call once to kick off game loop.
-	 * <p>
-	 * Runs as long as {@link #window} is not set to close.
-	 * <p>
-	 * In order, once per frame the {@link #frameTime} is calculated, input is processed, update code is run, the screen is rendered, and then also synced to the target framerate.
-	 */
 	private void loop() {
-		float elapsedTime;
-//		float accumulator = 0f;
-//		float interval = 1f / TARGET_UPS;
+		updateInterval = 1.0 / game.getTargetUps();
+		
+		double previousFrameTime = timer.getTime();
+		double loopTime = 0;
 		
 		while (!window.shouldClose()) {
-			// XXX Minor hack to limit frame updates when the window is "held" and instead tie pauses to at least the target UPS
-//			elapsedTime = Math.min(interval, timer.getElapsedTime());
-			elapsedTime = timer.getElapsedTime();
-			frameTime += elapsedTime;
+			profiler.startFrame();
+			double frameStartTime = timer.getTime();
+			double elapsedTime = (frameStartTime - previousFrameTime);
+			previousFrameTime = frameStartTime;
+			window.setFps((int) (1 / elapsedTime));
+			updateAccumulator += elapsedTime;
+			
+			loopTime += elapsedTime;
+			if (loopTime >= 1) {
+				profiler.printAverages();
+				loopTime = 0;
+			}
 
 			input();
 
-			update(elapsedTime);
+			update();
 			
 			render();
 			
-			// XXX Should syncing occur at the start of the frame? If a frame takes a long time to render/process, by the time we get here we could've already surpassed our frame time budget
-			sync();
+			profiler.endFrame();
+			// Get the time it took to actually process the entire frame (this should be counted as a segment of our overall time dictated by renderInterval)
+			double frameProcessTimeNanos = timer.getTime() - frameStartTime;
+			long sleepTime = (long) (Math.max(window.getTargetFps() - frameProcessTimeNanos, 0) * 1_000_000_000); // Convert to nanoseconds
+			
+			try {
+				sleepNanos(sleepTime);
+			} catch (InterruptedException e) {
+				AvoLog.log().warn("Sync operation was interrupted?", e);
+				Thread.currentThread().interrupt();
+			}
+			profiler.endBudget();
 		}
-		
-		// XXX Updates with a constant delta time, but blows up if frame rate is consistently low
-//		while (!window.shouldClose()) {
-//			// XXX Minor hack to limit frame updates when the window is "held" and instead tie pauses to at least the target UPS
-//			elapsedTime = Math.min(interval, timer.getElapsedTime());
-//			accumulator += elapsedTime;
-//			frameTime += elapsedTime;
-//
-//			input();
-//
-//			while (accumulator >= interval) {
-//				update(interval);
-//				accumulator -= interval;
-//			}
-//			
-//			render();
-//			
-//			// XXX Should syncing occur at the start of the frame? If a frame takes a long time to render/process, by the time we get here we could've already surpassed our frame time budget
-//			sync();
-//		}
 	}
 	
 	private void input() {
-		input.update();
+		profiler.inputStart();
+
+		window.pollEvents();
+		game.input(window);
+		
+		profiler.inputEnd();
+	}
+
+	private void update() {
+		profiler.updateStart();
+		
+		while (updateAccumulator > updateInterval) {
+			// TODO Process EventQueue
+			game.update((float) updateInterval);
+			updateAccumulator -= updateInterval;
+		}
+		
+		profiler.updateEnd();
 	}
 	
 	private void render() {
-		if (frameTime >= 1.0f) {
-			window.setFps(fps);
-//			logger.info("FPS: {}", fps);
-			frameTime = 0;
-			fps = 0;
-		}
-		fps++;
+		profiler.renderStart();
 		
-		game.render();
-		window.update();
+		game.render(window);
+		window.swapBuffers();
+		
+		profiler.renderEnd();
 	}
 	
-	private void update(float interval) {
-		// TODO Process EventQueue
-		game.update(interval);
-	}
-	
-	private void sync() {
-		float loopSlot = 1f / TARGET_FPS;
-		double endTime = timer.getLastLoopTime() + loopSlot + sleepRemainder;
-		
-		// Time values are seconds based, so we're subtracting 0.001 to make sure we won't try sleeping for fractions of milliseconds
-		while (timer.getTime() < endTime - 0.001) {
-			try {
+	/**
+	 * Perform regular {@link Thread#sleep()} while the remaining sleep duration is greater than {@code Avogine.SLEEP_PRECISION}.
+	 * If the remaining time is less than {@code Avogine.SLEEP_PRECISION} then perform {@link Thread#onSpinWait()} for the remainder of the time.
+	 * </p>
+	 * This method is preferable than just relying on {@code Thread.sleep()} since not only does {@code Thread.sleep()} not support sleeping for less
+	 * than a millisecond, but also because there's about a 0.5-3ms error rate on waking from the sleep so using a busy wait loop allows us to more 
+	 * quickly recover from smaller sleep durations.
+	 * 
+	 * @see <a href="http://andy-malakov.blogspot.com/2010/06/alternative-to-threadsleep.html">http://andy-malakov.blogspot.com/2010/06/alternative-to-threadsleep.html</a>
+	 * 
+	 * @param nanoDuration The amount of time to sleep for in nanoseconds.
+	 * @throws InterruptedException
+	 */
+	private static void sleepNanos(long nanoDuration) throws InterruptedException {
+		final long end = System.nanoTime() + nanoDuration;
+
+		long timeLeft = nanoDuration;
+		while (timeLeft > 0) {
+			if (timeLeft > SLEEP_PRECISION) {
 				Thread.sleep(1);
-			} catch (InterruptedException e) {
-				logger.warn("Sync operation was interrupted?", e);
-				Thread.currentThread().interrupt();
+			} else {
+				Thread.onSpinWait();
+			}
+			
+			timeLeft = end - System.nanoTime();
+			
+			if (Thread.interrupted()) {
+				throw new InterruptedException();
 			}
 		}
-		// If there is any time left over, save it for the next sync
-		// This ensures high precision when picking frame rate targets
-		sleepRemainder = (float) Math.max(0, endTime - timer.getTime());
 	}
 	
 	private void cleanup() {
