@@ -1,41 +1,28 @@
 package com.avogine.game;
 
+import org.lwjgl.glfw.*;
+
 /**
- *
+ * A wrapper class for managing calls to GLFW's time methods.
  */
 public class Timer {
 
-	private double lastLoopTime;
-
 	/**
-	 * 
+	 * Initialize the GLFW system timer.
+	 * </p>
+	 * This should largely be viewed as a convenience as the timer automatically gets initialized to 0
+	 * when {@link GLFW#glfwInit()} is called.
 	 */
 	public void init() {
-		lastLoopTime = getTime();
+		GLFW.glfwSetTime(0.0);
 	}
 
 	/**
-	 * @return
+	 * Returns the current time in seconds since {@link #init()} was called.
+	 * @return the current time in seconds since {@link #init()} was called.
 	 */
 	public double getTime() {
-		return System.nanoTime() / 1000_000_000.0;
+		return GLFW.glfwGetTime();
 	}
 
-	/**
-	 * @return
-	 */
-	public float getElapsedTime() {
-		double time = getTime();
-		float elapsedTime = (float) (time - lastLoopTime);
-		lastLoopTime = time;
-		return elapsedTime;
-	}
-
-	/**
-	 * @return
-	 */
-	public double getLastLoopTime() {
-		return lastLoopTime;
-	}
-	
 }
