@@ -1,85 +1,87 @@
 package com.avogine.io;
 
-import java.lang.invoke.*;
 import java.nio.*;
 import java.util.*;
 
-import org.lwjgl.*;
+import org.lwjgl.PointerBuffer;
 import org.lwjgl.glfw.*;
-import org.lwjgl.opengl.*;
+import org.lwjgl.opengl.GL11;
+import org.lwjgl.stb.STBImage;
 import org.lwjgl.system.*;
-import org.slf4j.*;
 
-import com.avogine.*;
-import com.avogine.experimental.annotation.*;
+import com.avogine.Avogine;
+import com.avogine.game.ui.nuklear.NuklearUI;
+import com.avogine.io.config.*;
+import com.avogine.logging.AvoLog;
+import com.avogine.util.ResourceUtils;
+import com.avogine.util.resource.ResourceConstants;
 
 /**
  * {@link Window} provides the primary entry point into OpenGL and GLFW.
  * <p>
  * A new instance of {@code Window} needs to be initialized before any GL or GLFW calls can be made. This
  * should be handled from {@link Avogine} as one of the required parameters to start is a {@code Window}.
- * {@code Window}s can also optionally be registered with an {@link Input} for custom input handling as
- * any system events like key presses or mouse motions are captured from the window context.
- * <p>
- * TODO WindowOptions
+ * {@code Window}s are also registered with an {@link Input} for custom input handling as any system events
+ * like key presses or mouse motions are captured from the window context.
+ * </p>
  */
 public class Window {
 
-	private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass().getSimpleName());
-
 	private long id;
+
+	private String title;
 	
 	private int width;
 	private int height;
-	private String title;
 	
+	private boolean fullscreen;
+	
+	private int maxFps;
+	private int maxBackgroundFps;
+	private double targetFPS;
 	private int fps;
 	
-	private WindowOptions options;
+	private boolean debugMode;
 	
-	public static boolean debugMode;
-		
+	private int monitorIndex;
 	private List<Long> monitorList = new ArrayList<>();
 	
-	private Input input;
+	private final Input input;
+	
+	private final List<NuklearUI> guiContexts;
 	
 	/**
-	 * @param width The width in pixels for this window
-	 * @param height The height in pixels for this window
 	 * @param title The window title
-	 * @param options window specific settings for configuring OpenGL
+	 * @param preferences 
 	 */
-	public Window(int width, int height, String title, WindowOptions options) {
-		this.width = width;
-		this.height = height;
+	public Window(String title, WindowPreferences preferences) {
 		this.title = title;
 		
-		this.options = options;
+		width = preferences.width();
+		height = preferences.height();
+		
+		fullscreen = preferences.fullscreen();
+		monitorIndex = preferences.monitor();
+		
+		maxFps = Math.clamp(preferences.fpsCap(), 0, 1000);
+		if (maxFps > 0) {
+			maxBackgroundFps = Math.clamp(preferences.backgroundFps(), 1, maxFps);
+		}
+		
+		input = new Input(this);
+		
+		guiContexts = new ArrayList<>();
 	}
 	
 	/**
-	 * 
+	 * @param windowConfig 
+	 * @param inputConfig 
 	 */
-	@InDev
-	public void init(Input input) {
-		GLFWErrorCallback.createPrint(System.err).set();
+	public void init(GLFWConfig windowConfig, InputConfig inputConfig) {
+		GLFW.glfwSetErrorCallback((int errorCode, long msgPtr) -> AvoLog.log().error("Error code [{}], msg [{}]", errorCode, MemoryUtil.memUTF8(msgPtr)));
 		
 		if (!GLFW.glfwInit()) {
 			throw new IllegalStateException("Could not initialize GLFW!");
-		}
-
-		GLFW.glfwWindowHint(GLFW.GLFW_CONTEXT_VERSION_MAJOR, 3);
-		GLFW.glfwWindowHint(GLFW.GLFW_CONTEXT_VERSION_MINOR, 3);
-		GLFW.glfwWindowHint(GLFW.GLFW_VISIBLE, GLFW.GLFW_FALSE);
-		GLFW.glfwWindowHint(GLFW.GLFW_RESIZABLE, GLFW.GLFW_TRUE);
-		GLFW.glfwWindowHint(GLFW.GLFW_SAMPLES, 4);
-//		GLFW.glfwWindowHint(GLFW.GLFW_DECORATED, GLFW.GLFW_FALSE);
-//		GLFW.glfwWindowHint(GLFW.GLFW_FLOATING, GLFW.GLFW_TRUE);
-//		GLFW.glfwWindowHint(GLFW.GLFW_TRANSPARENT_FRAMEBUFFER, GLFW.GLFW_TRUE);
-		
-		id = GLFW.glfwCreateWindow(width, height, title, MemoryUtil.NULL, MemoryUtil.NULL);
-		if (id == MemoryUtil.NULL) {
-			throw new RuntimeException("Failed to create window!");
 		}
 		
 		long monitor = GLFW.glfwGetPrimaryMonitor();
@@ -88,7 +90,20 @@ public class Window {
 			long monitorID = monitorBuffer.get();
 			monitorList.add(monitorID);
 		}
-		monitor = monitorList.get(options.monitorIndex);
+		monitor = monitorList.get(monitorIndex);
+		
+		windowConfig.windowHints().forEach(GLFW::glfwWindowHint);
+		
+		id = GLFW.glfwCreateWindow(width, height, title, fullscreen ? monitor : MemoryUtil.NULL, MemoryUtil.NULL);
+		if (id == MemoryUtil.NULL) {
+			throw new IllegalStateException("Failed to create window!");
+		}
+		
+		GLFW.glfwSetFramebufferSizeCallback(id, (window, w, h) -> {
+			width = w;
+			height = h;
+			GL11.glViewport(0, 0, width, height);
+		});
 		
 		try (MemoryStack stack = MemoryStack.stackPush()) {
 			IntBuffer pWidth = stack.mallocInt(1);
@@ -98,89 +113,38 @@ public class Window {
 			IntBuffer pY = stack.mallocInt(1);
 			
 			GLFW.glfwGetWindowSize(id, pWidth, pHeight);
-			
 			GLFW.glfwGetMonitorPos(monitor, pX, pY);
 			
 			GLFWVidMode videoMode = GLFW.glfwGetVideoMode(monitor);
 			
 			GLFW.glfwSetWindowPos(id, pX.get() + (videoMode.width() - pWidth.get(0)) / 2, pY.get() + (videoMode.height() - pHeight.get(0)) / 2);
 		}
-		
+
 		GLFW.glfwMakeContextCurrent(id);
 		
-		if (options.vsync) {
+		if (maxFps == 0) {
 			GLFW.glfwSwapInterval(1);
 		} else {
 			GLFW.glfwSwapInterval(0);
+			targetFPS = 1.0 / maxFps;
+			GLFW.glfwSetWindowFocusCallback(id, (window, focused) -> targetFPS = 1.0 / (focused ? maxFps : maxBackgroundFps));
 		}
-		
+
+		// XXX Customize by WindowConfig?
 		if (GLFW.glfwGetWindowAttrib(id, GLFW.GLFW_TRANSPARENT_FRAMEBUFFER) == GLFW.GLFW_TRUE) {
-			GLFW.glfwSetWindowOpacity(id, 0.5f);
+			GLFW.glfwSetWindowOpacity(id, 1.0f);
 		}
 		
-		GLFW.glfwShowWindow(id);
-
-		GL.createCapabilities();
+		loadAndSetIcons(ResourceConstants.TEXTURES.with("icon", "AGDG Logo.png"));
 		
-		// All of this should be handled through settings in WindowOptions
-		GL11.glClearColor(options.clearColor[0], options.clearColor[1], options.clearColor[2], options.clearColor[3]);
-		
-		if (options.depthTest) {
-			GL11.glEnable(GL11.GL_DEPTH_TEST);
-		}
-
-		if (options.blend) {
-			GL11.glEnable(GL11.GL_BLEND);
-			GL11.glBlendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA);
-		}
-		
-		if (options.cullFaces) {
-			GL11.glEnable(GL11.GL_CULL_FACE);
-			// TODO Fetch which face to cull from options as well
-			GL11.glCullFace(GL11.GL_BACK);
-		}
-
-		if (options.multisample) {
-			GL11.glEnable(GL13.GL_MULTISAMPLE);
-		}
-		
-		GLFW.glfwSetFramebufferSizeCallback(id, (window, w, h) -> {
-			width = w;
-			height = h;
-			GL11.glViewport(0, 0, width, height);
-			
-//			try (MemoryStack stack = MemoryStack.stackPush()) {
-//				FloatBuffer xscale = stack.mallocFloat(1);
-//				FloatBuffer yscale = stack.mallocFloat(1);
-//				
-//				GLFW.glfwGetWindowContentScale(id, xscale, yscale);
-//				
-//				logger.debug("{} {}", xscale.get(), yscale.get());
-//			}
-		});
-		
-		GLFW.glfwSetWindowPosCallback(id, (window, x, y) -> {
-			logger.debug("x: {} y: {}", x, y);
-		});
-		
+		// XXX Customizable by WindowConfig?
 		if (GLFW.glfwRawMouseMotionSupported()) {
 			GLFW.glfwSetInputMode(id, GLFW.GLFW_RAW_MOUSE_MOTION, GLFW.GLFW_TRUE);
 		}
+
+		GLFW.glfwShowWindow(id);
 		
-//		try (MemoryStack stack = MemoryStack.stackPush()) {
-//			FloatBuffer scaleX = stack.mallocFloat(1);
-//			FloatBuffer scaleY = stack.mallocFloat(1);
-//			
-//			GLFW.glfwGetMonitorContentScale(id, scaleX, scaleY);
-//			
-//			float contentScaleX = scaleX.get();
-//			float contentScaleY = scaleY.get();
-//			
-//			logger.debug("{} {}", contentScaleX, contentScaleY);
-//		}
-		
-		this.input = input;
-		input.init(this);
+		input.init(inputConfig);
 	}
 	
 	/**
@@ -189,23 +153,42 @@ public class Window {
 	 * Callers should not need to handle setting the OpenGL context before rendering as this method will automatically check
 	 * and, if needed, set this window context to be current.
 	 */
-	public void update() {
+	public void swapBuffers() {
 		if (GLFW.glfwGetCurrentContext() != id) {
 			GLFW.glfwMakeContextCurrent(id);
 		}
 		
 		GLFW.glfwSwapBuffers(id);
-		GLFW.glfwPollEvents();
 	}
 	
-	public void restoreState() {
-		GL11.glEnable(GL11.GL_DEPTH_TEST);
-		GL11.glEnable(GL11.GL_STENCIL_TEST);
-		GL11.glBlendFunc(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA);
-//		if (options.cullFace) {
-//			GL11.glEnable(GL11.GL_CULL_FACE);
-//			GL11.glCullFace(GL11.GL_BACK);
-//		}
+	/**
+	 * Poll GLFW for events and process them in {@link Input}.
+	 */
+	public void pollEvents() {
+		guiContexts.forEach(NuklearUI::inputBegin);
+		GLFW.glfwPollEvents();
+		guiContexts.forEach(NuklearUI::inputEnd);
+		
+		input.update();
+	}
+	
+	/**
+	 * Register a {@link NuklearUI} to this Window so that it's internal input state can be processed around calls to {@link #pollEvents()}.
+	 * @param gui The {@link NuklearUI} to register.
+	 * @return The {@link NuklearUI}.
+	 */
+	public NuklearUI registerGUIContext(NuklearUI gui) {
+		guiContexts.add(gui);
+		return gui;
+	}
+	
+	/**
+	 * Unregister a {@link NuklearUI} from this Window.
+	 * @param gui The {@link NuklearUI} to remove.
+	 * @return {@code true} if the {@link NuklearUI} was removed.
+	 */
+	public boolean removeGUIContext(NuklearUI gui) {
+		return guiContexts.remove(gui);
 	}
 	
 	/**
@@ -237,7 +220,39 @@ public class Window {
 		close();
 		
 		GLFW.glfwTerminate();
-		GLFW.glfwSetErrorCallback(null).free();
+		Objects.requireNonNull(GLFW.glfwSetErrorCallback(null)).free();
+	}
+	
+	/**
+	 * Read in image file data as {@link GLFWImage}s and load them as a Buffer to the Window icon.
+	 * @param filePaths a list of image file resource paths to load as application icons.
+	 */
+	private void loadAndSetIcons(String...filePaths) {
+		try (MemoryStack stack = MemoryStack.stackPush()) {
+			List<GLFWImage> iconList = new ArrayList<>();
+			
+			for (String filePath : filePaths) {
+				GLFWImage icon = GLFWImage.malloc(stack);
+
+				IntBuffer iconWidth = stack.mallocInt(1);
+				IntBuffer iconHeight = stack.mallocInt(1);
+				IntBuffer nrChannels = stack.mallocInt(1);
+
+				ByteBuffer fileData = ResourceUtils.readResourceToBuffer(filePath, 1024);
+				ByteBuffer imageData = STBImage.stbi_load_from_memory(fileData, iconWidth, iconHeight, nrChannels, 0);
+				if (imageData != null) {
+					icon.set(iconWidth.get(), iconHeight.get(), imageData);
+					iconList.add(icon);
+				} else {
+					AvoLog.log().warn("Icon failed to load: {}", filePath);
+				}
+			}
+			
+			GLFWImage.Buffer iconBuffer = GLFWImage.malloc(iconList.size(), stack);
+			iconList.forEach(iconBuffer::put);
+			iconBuffer.flip();
+			GLFW.glfwSetWindowIcon(id, iconBuffer);
+		}
 	}
 	
 	/**
@@ -247,14 +262,32 @@ public class Window {
 		return input;
 	}
 	
+	/**
+	 * Return target frames per second. Controls how often the screen is rendered in a single second.
+	 * </p>
+	 * @return target frames per second. Controls how often the screen is rendered in a single second.
+	 */
+	public double getTargetFps() {
+		return targetFPS;
+	}
+	
+	/**
+	 * @return An estimate value of how many frames are being rendered in this second based on elapsed time in the current game loop.
+	 */
 	public int getFps() {
 		return fps;
 	}
 	
+	/**
+	 * @param fps
+	 */
 	public void setFps(int fps) {
 		this.fps = fps;
 	}
 	
+	/**
+	 * @return The memory address of this {@link Window}.
+	 */
 	public long getId() {
 		return id;
 	}
@@ -282,35 +315,17 @@ public class Window {
 	}
 	
 	/**
-	 * Assorted options to be set prior to window creation to control optional parameters.
+	 * @param debugMode
 	 */
-	public static class WindowOptions {
-		
-		/**
-		 * Initialize the window in full screen mode.
-		 */
-		public boolean fullScreen;
-		
-		/**
-		 * Sync the window's refresh rate with the monitor's to avoid screen tearing.
-		 */
-		public boolean vsync;
-		
-		/**
-		 * The default monitor that the window should be placed on.
-		 */
-		public int monitorIndex = 0;
-		
-		public float[] clearColor = new float[] {100f / 255f, 149f / 255f, 237f / 255f, 1f};
-
-		public boolean depthTest;
-		
-		public boolean cullFaces;
-		
-		public boolean blend;
-		
-		public boolean multisample;
-		
+	public void setDebugMode(boolean debugMode) {
+		this.debugMode = debugMode;
+	}
+	
+	/**
+	 * @return the debugMode
+	 */
+	public boolean isDebugMode() {
+		return debugMode;
 	}
 	
 }
