@@ -1,265 +1,272 @@
 package com.avogine.render.loader.assimp;
 
-import java.lang.Math;
+import static org.lwjgl.system.MemoryUtil.*;
+
 import java.nio.*;
 import java.util.*;
 
 import org.joml.*;
-import org.lwjgl.*;
+import org.joml.Math;
 import org.lwjgl.assimp.*;
-import org.lwjgl.system.*;
+import org.lwjgl.system.MemoryStack;
 
+import com.avogine.render.TextureCache;
 import com.avogine.render.data.*;
-import com.avogine.render.data.mesh.*;
-import com.avogine.render.data.mesh.Mesh;
-import com.avogine.render.data.mesh.Texture.*;
-import com.avogine.render.loader.texture.*;
-import com.avogine.util.resource.*;
+import com.avogine.util.ResourceUtils;
 
 /**
  *
  */
 public class StaticModelLoader {
 	
-	private static final Map<String, Texture> textureCache = new HashMap<>();
-	
-	/**
-	 * Maximum number of diffuse textures to load from a model.
-	 * </p>
-	 * This is to facilitate a manageable shader uniform for activating specific texture material types.
-	 */
-	public static final int MAX_DIFFUSE_TEXTURES = 4;
-	/**
-	 * Maximum number of specular textures to load from a model.
-	 * </p>
-	 * This is to facilitate a manageable shader uniform for activating specific texture material types.
-	 */
-	public static final int MAX_SPECULAR_TEXTURES = 4;
-	
 	private StaticModelLoader() {
 		
 	}
 	
 	/**
-	 * @param resourcePath
-	 * @param texturesDir
+	 * @param id 
+	 * @param modelPath
+	 * @param textureCache 
 	 * @return
 	 */
-	public static Model load(String resourcePath, String texturesDir) {
-		//return load(resourcePath, texturesDir, Assimp.aiProcess_JoinIdenticalVertices | Assimp.aiProcess_Triangulate | Assimp.aiProcess_FixInfacingNormals, 1);
-		return load(resourcePath, texturesDir, Assimp.aiProcess_Triangulate, 1);
+	public static Model loadModel(String id, String modelPath, TextureCache textureCache) {
+		return loadModel(id, modelPath, textureCache, Assimp.aiProcess_GenSmoothNormals | Assimp.aiProcess_JoinIdenticalVertices |
+				Assimp.aiProcess_Triangulate | Assimp.aiProcess_FixInfacingNormals | Assimp.aiProcess_CalcTangentSpace | Assimp.aiProcess_LimitBoneWeights |
+				Assimp.aiProcess_GenBoundingBoxes);
 	}
 	
 	/**
-	 * @param resourcePath
-	 * @param texturesDir
+	 * @param id 
+	 * @param modelPath
+	 * @param textureCache 
 	 * @param flags
-	 * @param numberOfInstances
 	 * @return
 	 */
-	public static Model load(String resourcePath, String texturesDir, int flags, int numberOfInstances) {
-		AIFileIO fileIo = AIFileIO.create()
-	            .OpenProc((pFileIO, fileName, openMode) -> {
-	            	ByteBuffer data;
-	                String fileNameUtf8 = MemoryUtil.memUTF8(fileName);
-	                data = ResourceFileReader.ioResourceToByteBuffer(fileNameUtf8, 8 * 1024);
+	public static Model loadModel(String id, String modelPath, TextureCache textureCache, int flags) {
+		try (AIFileIO fileIo = AIFileIO.create()
+				.OpenProc((pFileIO, fileName, openMode) -> {
+					String fileNameUtf8 = memUTF8(fileName);
+					ByteBuffer data = ResourceUtils.readResourceToBuffer(fileNameUtf8);
 
-	                return AIFile.create()
-	                    .ReadProc((pFile, pBuffer, size, count) -> {
-	                        long max = Math.min(data.remaining(), size * count);
-	                        MemoryUtil.memCopy(MemoryUtil.memAddress(data) + data.position(), pBuffer, max);
-	                        return max;
-	                    })
-	                    .SeekProc((pFile, offset, origin) -> {
-	                        if (origin == Assimp.aiOrigin_CUR) {
-	                            data.position(data.position() + (int) offset);
-	                        } else if (origin == Assimp.aiOrigin_SET) {
-	                            data.position((int) offset);
-	                        } else if (origin == Assimp.aiOrigin_END) {
-	                            data.position(data.limit() + (int) offset);
-	                        }
-	                        return 0;
-	                    })
-	                    .FileSizeProc(pFile -> data.limit())
-	                    .address();
-	            })
-	            .CloseProc((pFileIO, pFile) -> {
-	                AIFile aiFile = AIFile.create(pFile);
+					return AIFile.create()
+							.ReadProc((pFile, pBuffer, size, count) -> {
+								long max = Math.min(data.remaining() / size, count);
+								memCopy(memAddress(data), pBuffer, max * size);
+								data.position(data.position() + (int) (max * size));
+								return max;
+							})
+							.SeekProc((pFile, offset, origin) -> {
+								if (origin == Assimp.aiOrigin_CUR) {
+									data.position(data.position() + (int) offset);
+								} else if (origin == Assimp.aiOrigin_SET) {
+									data.position((int) offset);
+								} else if (origin == Assimp.aiOrigin_END) {
+									data.position(data.limit() + (int) offset);
+								}
+								return 0;
+							})
+							.FileSizeProc(pFile -> data.limit())
+							.address();
+				})
+				.CloseProc((pFileIO, pFile) -> {
+					AIFile aiFile = AIFile.create(pFile);
 
-	                aiFile.ReadProc().free();
-	                aiFile.SeekProc().free();
-	                aiFile.FileSizeProc().free();
-	            });
-		
-		AIScene aiScene = Assimp.aiImportFileEx(ResourceConstants.MODEL_PATH + resourcePath, flags, fileIo);
-		
-		fileIo.OpenProc().free();
-		fileIo.CloseProc().free();
-		
-		if (aiScene == null) {
-			throw new RuntimeException("Error loading model: " + resourcePath);
+					aiFile.ReadProc().free();
+					aiFile.SeekProc().free();
+					aiFile.FileSizeProc().free();
+				});) {
+
+			AIScene aiScene = Assimp.aiImportFileEx(modelPath, flags, fileIo);
+
+			fileIo.OpenProc().free();
+			fileIo.CloseProc().free();
+
+			if (aiScene == null) {
+				throw new IllegalStateException(Assimp.aiGetErrorString());
+			}
+
+			String modelDirectory = modelPath.substring(0, modelPath.lastIndexOf('/') + 1);
+
+			int numMaterials = aiScene.mNumMaterials();
+			List<Material> materials = new ArrayList<>();
+			for (int i = 0; i < numMaterials; ++i) {
+				AIMaterial aiMaterial = AIMaterial.create(aiScene.mMaterials().get(i));
+				materials.add(processMaterial(aiMaterial, modelDirectory, textureCache));
+			}
+
+			List<Mesh> meshes = processNode(aiScene.mRootNode(), aiScene);
+
+			Assimp.aiReleaseImport(aiScene);
+
+			return new Model(id, meshes, materials);
 		}
+	}
 
-		int numMaterials = aiScene.mNumMaterials();
-		PointerBuffer aiMaterials = aiScene.mMaterials();
-		List<Material> materials = new ArrayList<>();
-		for (int i = 0; i < numMaterials; i++) {
-			AIMaterial aiMaterial = AIMaterial.create(aiMaterials.get(i));
-			processMaterial(aiMaterial, materials);
+	/**
+	 * <a href="https://github.com/Avoduhcado/MAvogine/issues/41">PBR materials #41</a>
+	 */
+	protected static Material processMaterial(AIMaterial aiMaterial, String modelDirectory, TextureCache textureCache) {
+		// TODO#41 Use a real default
+		Material material = new Material(textureCache.getDefaultTexture());
+		try (MemoryStack stack = MemoryStack.stackPush()) {
+			AIColor4D color = AIColor4D.create();
+	
+			int result = Assimp.aiGetMaterialColor(aiMaterial, Assimp.AI_MATKEY_COLOR_DIFFUSE, Assimp.aiTextureType_NONE, 0, color);
+			if (result == Assimp.aiReturn_SUCCESS) {
+				material.setDiffuseColor(new Vector4f(color.r(), color.g(), color.b(), color.a()));
+			}
+	
+			result = Assimp.aiGetMaterialColor(aiMaterial, Assimp.AI_MATKEY_COLOR_AMBIENT, Assimp.aiTextureType_NONE, 0, color);
+			if (result == Assimp.aiReturn_SUCCESS) {
+				material.setAmbientColor(new Vector4f(color.r(), color.g(), color.b(), color.a()));
+			}
+			
+			result = Assimp.aiGetMaterialColor(aiMaterial, Assimp.AI_MATKEY_COLOR_SPECULAR, Assimp.aiTextureType_NONE, 0, color);
+			if (result == Assimp.aiReturn_SUCCESS) {
+				material.setSpecularColor(new Vector4f(color.r(), color.g(), color.b(), color.a()));
+			}
+
+			// TODO#41 Add reflectance to Material
+//			float reflectance = 0.0f;
+//			float[] shininessFactor = new float[] { 0.0f };
+//			int[] pMax = new int[] { 1 };
+//			result = Assimp.aiGetMaterialFloatArray(aiMaterial, Assimp.AI_MATKEY_SHININESS_STRENGTH, Assimp.aiTextureType_NONE, 0, shininessFactor, pMax);
+//			if (result != Assimp.aiReturn_SUCCESS) {
+//				reflectance = shininessFactor[0];
+//			}
+//			material.setReflectance(reflectance);
+			
+			AIString aiTexturePath = AIString.calloc(stack);
+			Assimp.aiGetMaterialTexture(aiMaterial, Assimp.aiTextureType_DIFFUSE, 0, aiTexturePath, (IntBuffer) null, null, null, null, null, null);
+			String texturePath = aiTexturePath.dataString();
+			if (texturePath != null && !texturePath.isBlank()) {
+				material.setDiffuseTexture(textureCache.getTexture(modelDirectory + texturePath));
+				material.setDiffuseColor(Material.DEFAULT_COLOR);
+			}
+			
+			// TODO#41 Add normal maps and potentially ambient/specular mapping
+			
+			return material;
 		}
-		
-		List<Mesh> meshes = new ArrayList<>();
-		processNode(aiScene.mRootNode(), aiScene, meshes, materials, texturesDir);
-
-		Assimp.aiReleaseImport(aiScene);
-
-		var model = new Model(meshes, materials);
-
-		return model;
 	}
 	
-	protected static void processNode(AINode node, AIScene scene, List<Mesh> meshes, List<Material> materials, String texturesDir) {
+	protected static List<Mesh> processNode(AINode node, AIScene scene) {
+		List<Mesh> meshes = new ArrayList<>();
 		// process all the node's meshes (if any)
 		for (int i = 0; i < node.mNumMeshes(); i++) {
 			AIMesh mesh = AIMesh.create(scene.mMeshes().get(node.mMeshes().get(i)));
-			meshes.add(processMesh(mesh, scene, materials, texturesDir));
+			meshes.add(processMesh(mesh));
 		}
 		
 		// then do the same for each of its children
 		for (int i = 0; i < node.mNumChildren(); i++) {
-			processNode(AINode.create(node.mChildren().get(i)), scene, meshes, materials, texturesDir);
+			meshes.addAll(processNode(AINode.create(node.mChildren().get(i)), scene));
 		}
+		return meshes;
 	}
 	
-	protected static Mesh processMesh(AIMesh aiMesh, AIScene scene, List<Material> materials, String texturesDir) {
-		List<Vertex> vertices = processVertices(aiMesh);
-		List<Integer> indices = processIndices(aiMesh);
-		List<Texture> textures = processTextures(aiMesh, scene, texturesDir);
-		
-//		Material material = new Material();
-		int materialIdx = aiMesh.mMaterialIndex();
-		// TODO Store material cache to load from?
-//		if (materialIdx >= 0 && materialIdx < materials.size()) {
-//			material = materials.get(materialIdx);
-//		}
-//		mesh.material = material;
+	protected static Mesh processMesh(AIMesh aiMesh) {
+		var vertexData = new VertexData(processVertices(aiMesh), processNormals(aiMesh), processTangents(aiMesh), processBitangents(aiMesh), processTextureCoordinates(aiMesh), processIndices(aiMesh));
+		int materialIndex = aiMesh.mMaterialIndex();
+		AIAABB aabb = aiMesh.mAABB();
+		Vector3f aabbMin = new Vector3f(aabb.mMin().x(), aabb.mMin().y(), aabb.mMin().z());
+		Vector3f aabbMax = new Vector3f(aabb.mMax().x(), aabb.mMax().y(), aabb.mMax().z());
 
-		Mesh mesh = new Mesh(vertices, indices, textures, materialIdx);
-
-		return mesh;
+		return new Mesh(vertexData, materialIndex, aabbMin, aabbMax);
 	}
 	
-	protected static List<Texture> processTextures(AIMesh mesh, AIScene scene, String texturesDir) {
-		AIMaterial aiMaterial = AIMaterial.create(scene.mMaterials().get(mesh.mMaterialIndex()));
-		List<Texture> textures = new ArrayList<>();
-		
-		int diffuseCount = Assimp.aiGetMaterialTextureCount(aiMaterial, Assimp.aiTextureType_DIFFUSE);
-		for (int i = 0; i < Math.min(diffuseCount, MAX_DIFFUSE_TEXTURES); i++) {
-			textures.add(processTexture(aiMaterial, texturesDir, TextureType.DIFFUSE, Assimp.aiTextureType_DIFFUSE, i));
+	private static float[] processVertices(AIMesh aiMesh) {
+		AIVector3D.Buffer buffer = aiMesh.mVertices();
+		float[] data = new float[buffer.remaining() * 3];
+		int pos = 0;
+		while (buffer.hasRemaining()) {
+			AIVector3D vertex = buffer.get();
+			data[pos++] = vertex.x();
+			data[pos++] = vertex.y();
+			data[pos++] = vertex.z();
 		}
-		int specularCount = Assimp.aiGetMaterialTextureCount(aiMaterial, Assimp.aiTextureType_SPECULAR);
-		for (int i = 0; i < Math.min(specularCount, MAX_SPECULAR_TEXTURES); i++) {
-			textures.add(processTexture(aiMaterial, texturesDir, TextureType.SPECULAR, Assimp.aiTextureType_SPECULAR, i));
-		}
-		
-		return textures;
-	}
-
-	protected static Texture processTexture(AIMaterial aiMaterial, String texturesDir, TextureType textureType, int assimpTextureType, int index) {
-		AIString path = AIString.calloc();
-		Assimp.aiGetMaterialTexture(aiMaterial, assimpTextureType, index, path, (IntBuffer) null, null, null, null, null, null);
-		String textPath = path.dataString();
-		Texture texture = null;
-		
-		if (textPath != null && !textPath.isBlank()) {
-			String textureFile = "";
-			if (texturesDir != null && !texturesDir.isBlank()) {
-				textureFile += texturesDir + "/";
-			}
-			textureFile += textPath;
-			textureFile = textureFile.replace("//", "/");
-			// TODO Reimplement texture cache with "new" Texture class
-			texture = textureCache.computeIfAbsent(textureFile, TextureLoader::loadTexturePro);
-			texture.setType(textureType);
-		}
-		return texture;
+		return data;
 	}
 	
-	protected static void processMaterial(AIMaterial aiMaterial, List<Material> materials) {
-		AIColor4D color = AIColor4D.create();
-		
-		Vector3f ambient = Material.DEFAULT_COLOR;
-		int result = Assimp.aiGetMaterialColor(aiMaterial, Assimp.AI_MATKEY_COLOR_AMBIENT, Assimp.aiTextureType_NONE, 0, color);
-		if (result == 0) {
-			ambient = new Vector3f(color.r(), color.g(), color.b());
+	private static float[] processNormals(AIMesh aiMesh) {
+		if (aiMesh.isNull(AIMesh.MNORMALS)) {
+			return new float[aiMesh.mNumVertices() * 3];
 		}
 		
-		Vector3f diffuse = Material.DEFAULT_COLOR;
-		result = Assimp.aiGetMaterialColor(aiMaterial, Assimp.AI_MATKEY_COLOR_DIFFUSE, Assimp.aiTextureType_NONE, 0, color);
-		if (result == 0) {
-			diffuse = new Vector3f(color.r(), color.g(), color.b());
+		AIVector3D.Buffer buffer = aiMesh.mNormals();
+		float[] data = new float[buffer.remaining() * 3];
+		int pos = 0;
+		while (buffer.hasRemaining()) {
+			AIVector3D normal = buffer.get();
+			data[pos++] = normal.x();
+			data[pos++] = normal.y();
+			data[pos++] = normal.z();
 		}
-
-		Vector3f specular = Material.DEFAULT_COLOR;
-		result = Assimp.aiGetMaterialColor(aiMaterial, Assimp.AI_MATKEY_COLOR_SPECULAR, Assimp.aiTextureType_NONE, 0, color);
-		if (result == 0) {
-			specular = new Vector3f(color.r(), color.g(), color.b());
-		}
-
-		Material material = new Material(ambient, diffuse, specular, 32f);
-		materials.add(material);
-	}
-
-	private static List<Vertex> processVertices(AIMesh aiMesh) {
-		AIVector3D.Buffer aiPositions = aiMesh.mVertices();
-		AIVector3D.Buffer aiNormals = aiMesh.mNormals();
-		// XXX Potentially support multiple texture coordinates per mesh?
-		AIVector3D.Buffer aiTextureCoordinates = aiMesh.mTextureCoords().hasRemaining() ? aiMesh.mTextureCoords(0) : null;
-
-		List<Vertex> vertices = new ArrayList<>();
-		
-		Vector3f position = new Vector3f();
-		Vector3f normal = new Vector3f();
-		Vector2f textureCoordinate = new Vector2f();
-		
-		for (int i = 0; i < aiMesh.mNumVertices(); i++) {
-			AIVector3D aiPosition = aiPositions.get();
-			position.set(aiPosition.x(), aiPosition.y(), aiPosition.z());
-			
-			AIVector3D aiNormal = aiNormals.get();
-			normal.set(aiNormal.x(), aiNormal.y(), aiNormal.z());
-			
-			textureCoordinate.set(0);
-			if (aiTextureCoordinates != null)  {
-				AIVector3D aiTextureCoordinate = aiTextureCoordinates.get();
-				textureCoordinate.set(aiTextureCoordinate.x(), aiTextureCoordinate.y());
-			}
-			
-			Vertex vertex = new Vertex(position.x, position.y, position.z,
-					normal.x, normal.y, normal.z,
-					textureCoordinate.x, textureCoordinate.y);
-			vertices.add(vertex);
-		}
-
-		return vertices;
+		return data;
 	}
 	
-	private static List<Integer> processIndices(AIMesh aiMesh) {
-		// As long as the aiProcess_Triangulate flag is being used we should be good to assume that each face is a triangle
+	private static float[] processTangents(AIMesh aiMesh) {
+		if (aiMesh.isNull(AIMesh.MTANGENTS)) {
+			return new float[aiMesh.mNumVertices() * 3];
+		}
+		
+		AIVector3D.Buffer buffer = aiMesh.mTangents();
+		float[] data = new float[buffer.remaining() * 3];
+		int pos = 0;
+		while (buffer.hasRemaining()) {
+			AIVector3D tangent = buffer.get();
+			data[pos++] = tangent.x();
+			data[pos++] = tangent.y();
+			data[pos++] = tangent.z();
+		}
+		return data;
+	}
+	
+	private static float[] processBitangents(AIMesh aiMesh) {
+		if (aiMesh.isNull(AIMesh.MBITANGENTS)) {
+			return new float[aiMesh.mNumVertices() * 3];
+		}
+		
+		AIVector3D.Buffer buffer = aiMesh.mBitangents();
+		float[] data = new float[buffer.remaining() * 3];
+		int pos = 0;
+		while (buffer.hasRemaining()) {
+			AIVector3D bitangent = buffer.get();
+			data[pos++] = bitangent.x();
+			data[pos++] = bitangent.y();
+			data[pos++] = bitangent.z();
+		}
+		return data;
+	}
+	
+	private static float[] processTextureCoordinates(AIMesh aiMesh) {
+		if (aiMesh.isNull(AIMesh.MTEXTURECOORDS)) {
+			return new float[aiMesh.mNumVertices() * 2];
+		}
+		
+		AIVector3D.Buffer buffer = aiMesh.mTextureCoords(0);
+		float[] data = new float[buffer.remaining() * 2];
+		int pos = 0;
+		while (buffer.hasRemaining()) {
+			AIVector3D textureCoordinate = buffer.get();
+			data[pos++] = textureCoordinate.x();
+			data[pos++] = textureCoordinate.y();
+		}
+		return data;
+	}
+	
+	private static int[] processIndices(AIMesh aiMesh) {
+		List<Integer> indices = new ArrayList<>();
 		int numFaces = aiMesh.mNumFaces();
 		AIFace.Buffer aiFaces = aiMesh.mFaces();
-		
-		List<Integer> indices = new ArrayList<>();
-
 		for (int i = 0; i < numFaces; i++) {
 			AIFace aiFace = aiFaces.get(i);
 			IntBuffer buffer = aiFace.mIndices();
-			while (buffer.remaining() > 0) {
+			while (buffer.hasRemaining()) {
 				indices.add(buffer.get());
 			}
 		}
-
-		return indices;
+		return indices.stream().mapToInt(Integer::intValue).toArray();
 	}
 	
 }
