@@ -1,9 +1,8 @@
 package com.avogine.render.loader.parshapes;
 
 import org.lwjgl.util.par.ParShapes;
-import org.lwjgl.util.par.ParShapesMesh;
 
-import com.avogine.render.data.*;
+import com.avogine.render.data.Mesh;
 
 /**
  * TODO Add param objects to pass to each method for shape customization
@@ -33,80 +32,80 @@ public class ParShapesLoader {
 	/**
 	 * @return
 	 */
-	public static RawMesh loadPlane() {
-		ParShapesMesh parMesh = ParShapes.par_shapes_create_plane(100, 100);
-		ParShapes.par_shapes_scale(parMesh, 512, 512, 1);
-		ParShapes.par_shapes_rotate(parMesh, (float) Math.toRadians(-90), new float[] {1, 0, 0});
-		ParShapes.par_shapes_translate(parMesh, -256, 0, 256);
-		
-		RawMesh mesh = new RawMesh(parMesh.points(parMesh.npoints() * 3));
-		mesh.addAttribute(1, parMesh.tcoords(parMesh.npoints() * 2), 2);
-		mesh.addAttribute(2, parMesh.normals(parMesh.npoints() * 3), 3);
-		mesh.addIndexAttribute(parMesh.triangles(parMesh.ntriangles() * 3), parMesh.ntriangles() * 3);
-		
-		parMesh.free();
-		
-		return mesh;
-	}
-	
-	/**
-	 * @param radius
-	 * @return
-	 */
-	public static RawMesh loadSphere(float radius) {
-		ParShapesMesh parMesh = ParShapes.par_shapes_create_parametric_sphere(16, 16);
-		
-		ParShapes.par_shapes_scale(parMesh, radius, radius, radius);
-		ParShapes.par_shapes_translate(parMesh, 0, radius * 0.5f, 0);
-		
-		RawMesh mesh = new RawMesh(parMesh.points(parMesh.npoints() * 3));
-		mesh.addAttribute(1, parMesh.tcoords(parMesh.npoints() * 2), 2);
-		mesh.addAttribute(2, parMesh.normals(parMesh.npoints() * 3), 3);
-		mesh.addIndexAttribute(parMesh.triangles(parMesh.ntriangles() * 3), parMesh.ntriangles() * 3);
-		
-		parMesh.free();
-		
-		return mesh;
-	}
-	
-	/**
-	 * @return
-	 */
-	public static RawMesh loadLSystem() {
-		String program =
-	            "sx 2 sy 2" +
-	            " ry -90 rx 90" +
-	            " shape tube rx 15  call rlimb rx -15" +
-	            " shape tube rx -15 call llimb rx 15" +
-	            " shape tube ry 15  call rlimb ry -15" +
-	            " shape tube ry 15  call llimb ry -15" +
-	            " rule rlimb" +
-	            "     sx 0.925 sy 0.925 tz 1 rx 1.2" +
-	            "     call rlimb2" +
-	            " rule rlimb2.1" +
-	            "     shape connect" +
-	            "     call rlimb" +
-	            " rule rlimb2.1" +
-	            "     rx 15  shape tube call rlimb rx -15" +
-	            "     rx -15 shape tube call llimb rx 15" +
-	            " rule rlimb.1" +
-	            "     call llimb" +
-	            " rule llimb.1" +
-	            "     call rlimb" +
-	            " rule llimb.10" +
-	            "     sx 0.925 sy 0.925" +
-	            "     tz 1" +
-	            "     rx -1.2" +
-	            "     shape connect" +
-	            "     call llimb";
-		ParShapesMesh parMesh = ParShapes.par_shapes_create_lsystem(program, 5, 60);
-		
-		RawMesh mesh = new RawMesh(parMesh.points(parMesh.npoints() * 3));
-		mesh.addIndexAttribute(parMesh.triangles(parMesh.ntriangles() * 3), parMesh.ntriangles() * 3);
-		
-		parMesh.free();
-		
-		return mesh;
-	}
+//	public static RawMesh loadPlane() {
+//		ParShapesMesh parMesh = ParShapes.par_shapes_create_plane(100, 100);
+//		ParShapes.par_shapes_scale(parMesh, 512, 512, 1);
+//		ParShapes.par_shapes_rotate(parMesh, (float) Math.toRadians(-90), new float[] {1, 0, 0});
+//		ParShapes.par_shapes_translate(parMesh, -256, 0, 256);
+//		
+//		RawMesh mesh = new RawMesh(parMesh.points(parMesh.npoints() * 3));
+//		mesh.addAttribute(1, parMesh.tcoords(parMesh.npoints() * 2), 2);
+//		mesh.addAttribute(2, parMesh.normals(parMesh.npoints() * 3), 3);
+//		mesh.addIndexAttribute(parMesh.triangles(parMesh.ntriangles() * 3), parMesh.ntriangles() * 3);
+//		
+//		parMesh.free();
+//		
+//		return mesh;
+//	}
+//	
+//	/**
+//	 * @param radius
+//	 * @return
+//	 */
+//	public static RawMesh loadSphere(float radius) {
+//		ParShapesMesh parMesh = ParShapes.par_shapes_create_parametric_sphere(16, 16);
+//		
+//		ParShapes.par_shapes_scale(parMesh, radius, radius, radius);
+//		ParShapes.par_shapes_translate(parMesh, 0, radius * 0.5f, 0);
+//		
+//		RawMesh mesh = new RawMesh(parMesh.points(parMesh.npoints() * 3));
+//		mesh.addAttribute(1, parMesh.tcoords(parMesh.npoints() * 2), 2);
+//		mesh.addAttribute(2, parMesh.normals(parMesh.npoints() * 3), 3);
+//		mesh.addIndexAttribute(parMesh.triangles(parMesh.ntriangles() * 3), parMesh.ntriangles() * 3);
+//		
+//		parMesh.free();
+//		
+//		return mesh;
+//	}
+//	
+//	/**
+//	 * @return
+//	 */
+//	public static RawMesh loadLSystem() {
+//		String program =
+//	            "sx 2 sy 2" +
+//	            " ry -90 rx 90" +
+//	            " shape tube rx 15  call rlimb rx -15" +
+//	            " shape tube rx -15 call llimb rx 15" +
+//	            " shape tube ry 15  call rlimb ry -15" +
+//	            " shape tube ry 15  call llimb ry -15" +
+//	            " rule rlimb" +
+//	            "     sx 0.925 sy 0.925 tz 1 rx 1.2" +
+//	            "     call rlimb2" +
+//	            " rule rlimb2.1" +
+//	            "     shape connect" +
+//	            "     call rlimb" +
+//	            " rule rlimb2.1" +
+//	            "     rx 15  shape tube call rlimb rx -15" +
+//	            "     rx -15 shape tube call llimb rx 15" +
+//	            " rule rlimb.1" +
+//	            "     call llimb" +
+//	            " rule llimb.1" +
+//	            "     call rlimb" +
+//	            " rule llimb.10" +
+//	            "     sx 0.925 sy 0.925" +
+//	            "     tz 1" +
+//	            "     rx -1.2" +
+//	            "     shape connect" +
+//	            "     call llimb";
+//		ParShapesMesh parMesh = ParShapes.par_shapes_create_lsystem(program, 5, 60, null, 0);
+//		
+//		RawMesh mesh = new RawMesh(parMesh.points(parMesh.npoints() * 3));
+//		mesh.addIndexAttribute(parMesh.triangles(parMesh.ntriangles() * 3), parMesh.ntriangles() * 3);
+//		
+//		parMesh.free();
+//		
+//		return mesh;
+//	}
 	
 }
