@@ -1,7 +1,6 @@
 package com.avogine.render.shader;
 
-import com.avogine.render.shader.uniform.UniformMat4;
-import com.avogine.render.shader.uniform.UniformSampler;
+import com.avogine.render.shader.uniform.*;
 
 /**
  *
@@ -11,7 +10,8 @@ public class FontShader extends ShaderProgram {
 	public UniformMat4 projection = new UniformMat4();
 	public UniformMat4 model = new UniformMat4();
 	
-	public UniformSampler fontSample = new UniformSampler();
+	public UniformSampler fontTexture = new UniformSampler();
+	public UniformVec4 textColor = new UniformVec4();
 	
 	/**
 	 * @param vertexShaderFile
@@ -19,13 +19,13 @@ public class FontShader extends ShaderProgram {
 	 */
 	public FontShader(String vertexShaderFile, String fragmentShaderFile) {
 		super(vertexShaderFile, fragmentShaderFile);
-		storeAllUniformLocations(projection, model, fontSample);
+		storeAllUniformLocations(projection, model, fontTexture, textColor);
 		loadTexUnit();
 	}
 	
 	private void loadTexUnit() {
 		bind();
-		fontSample.loadTexUnit(0);
+		fontTexture.loadTexUnit(0);
 		unbind();
 	}
 
