@@ -2,14 +2,23 @@ package com.avogine.render.shader.uniform;
 
 import org.lwjgl.opengl.GL20;
 
+import com.avogine.logging.AvoLog;
+
+/**
+ * Base class for uniform variables to be passed along to OpenGL shaders.
+ */
 public abstract class Uniform {
 
 	private int location;
 	
+	/**
+	 * @param programID
+	 * @param name
+	 */
 	public void storeUniformLocation(int programID, String name) {
 		location = GL20.glGetUniformLocation(programID, name);
 		if (location < 0) {
-			System.err.println("No uniform variable called " + name + " found!");
+			AvoLog.log().warn("No uniform variable called {} found!", name);
 		}
 	}
 	
