@@ -1,6 +1,6 @@
 package com.avogine.render.shader.uniform;
 
-import com.avogine.render.data.*;
+import com.avogine.render.data.Material;
 
 /**
  *
@@ -22,8 +22,8 @@ public class UniformMaterial extends Uniform {
 	 * @param material
 	 */
 	public void loadMaterial(Material material) {
-		// TODO Do we want to handle loading the tex unit here, or handle activating/binding textures here? Or neither
-		shininess.loadFloat(material.getShininess());
+		// TODO#41 Do we want to handle loading the tex unit here, or handle activating/binding textures here? Or neither
+//		shininess.loadFloat(material.shininess());
 	}
 	
 }
