@@ -1,29 +1,32 @@
 package com.avogine.util;
 
+/**
+ * Utility class for managing arrays.
+ */
 public class ArrayUtils {
 
-	public static final float[] EMPTY_FLOAT_ARRAY = new float[0];
-	
+	private static final float[] EMPTY_FLOAT_ARRAY = new float[0];
+
 	/**
-     * <p>Converts an array of object Floats to primitives.
-     *
-     * <p>This method returns {@code null} for a {@code null} input array.
-     *
-     * @param array  a {@code Float} array, may be {@code null}
-     * @return a {@code float} array, {@code null} if null array input
-     * @throws NullPointerException if array content is {@code null}
-     */
-    public static float[] toPrimitive(final Float[] array) {
-        if (array == null) {
-            return null;
-        } else if (array.length == 0) {
-            return EMPTY_FLOAT_ARRAY;
-        }
-        final float[] result = new float[array.length];
-        for (int i = 0; i < array.length; i++) {
-            result[i] = array[i].floatValue();
-        }
-        return result;
-    }
-	
+	 * <p>Converts an array of object Floats to primitives.
+	 *
+	 * <p>This method returns {@code null} for a {@code null} input array.
+	 *
+	 * @param array  a {@code Float} array, may be {@code null}
+	 * @return a {@code float} array, {@code null} if null array input
+	 * @throws NullPointerException if array content is {@code null}
+	 */
+	public static float[] toPrimitive(final Float[] array) {
+		if (array == null) {
+			return null;
+		} else if (array.length == 0) {
+			return EMPTY_FLOAT_ARRAY;
+		}
+		final float[] result = new float[array.length];
+		for (int i = 0; i < array.length; i++) {
+			result[i] = array[i].floatValue();
+		}
+		return result;
+	}
+
 }
