@@ -1,15 +1,16 @@
 package com.avogine.serialize;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.*;
 
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.Test;
 
-import com.avogine.game.scene.*;
-import com.fasterxml.jackson.annotation.*;
-import com.fasterxml.jackson.annotation.JsonAutoDetect.*;
-import com.fasterxml.jackson.core.*;
+import com.avogine.game.scene.Scene;
+import com.avogine.io.serializer.SceneMapper;
+import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
+import com.fasterxml.jackson.annotation.PropertyAccessor;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.*;
 
 /**
@@ -20,7 +21,7 @@ class SceneMapperTest {
 	@Test
 	void serializeTest() throws JsonMappingException, JsonProcessingException {
 		var mapper = new ObjectMapper();
-	    mapper.setVisibility(PropertyAccessor.FIELD, Visibility.ANY);
+		mapper.setVisibility(PropertyAccessor.FIELD, Visibility.ANY);
 //		mapper.disable(SerializationFeature.FAIL_ON_EMPTY_BEANS);
 		
 		Car car = new Car("Yellow", 1994);
@@ -39,6 +40,12 @@ class SceneMapperTest {
 		assertEquals(1994, jsonNode.get("year").asInt());
 	}
 	
+	/**
+	 * TODO Make a proper test for scene serialization support, pass in a target file to serialize too and use Junit temp files to handle automatically cleaning
+	 * @see SceneMapper#serializeScene(Scene)
+	 * @param scene
+	 * @throws IOException
+	 */
 	public static void serializeScene(Scene scene) throws IOException {
 		var mapper = new ObjectMapper();
 //	    mapper.setVisibility(PropertyAccessor.FIELD, Visibility.ANY);
